Move version-specific CLI code into _juju*.py.
