+++ txjuju/_juju2.py
@@ -4,6 +4,8 @@ import os.path
 
 import yaml
 
+from . import _utils
+
 
 class ConfigWriter(object):
     """The JujuConfig writer specific to Juju 2.x."""
@@ -107,3 +109,62 @@ class ConfigWriter(object):
                 # but currently fails on juju2beta6.
                 pass
             bootstraps[controller.name] = config
+
+
+class CLIHooks(object):
+
+    CFGDIR_ENVVAR = "JUJU_DATA"
+
+    def get_bootstrap_args(self, spec, to=None, cfgfile=None,
+                           verbose=False, gui=False, autoupgrade=False):
+        args = ["bootstrap"]
+        if verbose:
+            args += ["-v"]
+        if to:
+            args += ["--to", to]
+        if not autoupgrade:
+            args += ["--no-auto-upgrade"]
+        if not gui:
+            args += ["--no-gui"]
+        if cfgfile:
+            args += ["--config", cfgfile]
+        args += [spec.name, spec.driver]
+        return args
+
+    def get_api_info_args(self, controller_name=None):
+        args = ["show-controller", "--show-password", "--format=yaml"]
+        if controller_name is not None:
+            args += [controller_name]
+        return args
+
+    def parse_api_info(self, output, controller_name=None):
+        data = yaml.load(output, _utils.UnicodeYamlLoader)
+        if controller_name is None:
+            if len(data) > 1:
+                raise RuntimeError(
+                    "got back {} results, expected 1".format(len(data)))
+            _, data = data.popitem()
+        else:
+            data = data[controller_name]
+
+        info = {
+            "endpoints": data["details"]["api-endpoints"],
+            "user": data["account"]["user"].rstrip("@local"),
+            "password": data["account"]["password"],
+            "model_uuid": None,  # This will be set below for each model.
+            }
+        infos = {None: info}  # Start with the non-model API info.
+        for modelname, modelinfo in data["models"].items():
+            # Strip off the "<user>@local/" part.
+            modelname = modelname.rpartition("/")[2]
+            infos[modelname] = dict(info, model_uuid=modelinfo["uuid"])
+        return infos
+
+    def get_destroy_controller_args(self, name=None, force=False):
+        if force:
+            args = ["kill-controller", "--yes"]
+        else:
+            args = ["destroy-controller", "--yes", "--destroy-all-models"]
+        if name is not None:
+            args += [name]
+        return args
